memory-hotplug: bound device config sizes to u32
The previous commit bounded requested_size_mib, but the sizes accepted
by PUT /hotplug/memory were still usize. mib_to_bytes shifts left by 20
and discards the bits shifted out, so a large enough total_size_mib
wrapped to a smaller byte count: 2^44 MiB became a 0-byte region.

Use u32 for total_size_mib, block_size_mib and slot_size_mib,
matching requested_size_mib and the balloon's amount_mib. Since
u32::MAX MiB is just under 4 PiB, every accepted value fits a u64 byte
count, so the wrap becomes unrepresentable at deserialization rather
than something to check for later.

Route the conversions through u32_mib_to_bytes and bytes_to_u32_mib,
which widen before shifting, so the MiB conversion is no longer
open-coded.

Document the ranges in the API spec and record the behaviour change in
the changelog.

+++ src/vmm/src/devices/virtio/mem/mod.rs
@@ -16,7 +16,7 @@ pub(crate) const MEM_NUM_QUEUES: usize = 1;
 
 pub(crate) const MEM_QUEUE: usize = 0;
 
-pub const VIRTIO_MEM_DEFAULT_BLOCK_SIZE_MIB: usize = 2;
-pub const VIRTIO_MEM_DEFAULT_SLOT_SIZE_MIB: usize = 128;
+pub const VIRTIO_MEM_DEFAULT_BLOCK_SIZE_MIB: u32 = 2;
+pub const VIRTIO_MEM_DEFAULT_SLOT_SIZE_MIB: u32 = 128;
 
 pub const VIRTIO_MEM_DEV_ID: &str = "mem";
